Added a Helptext, added another Function
For more overview for the future User of the Programm, is started
to use an Help-Text. This one will be regualry updated to guiede
the user through the programm. The call comes through the function
help() mentioned in input.c. As well i added a function called
ExecuteStructure, which refers to the Chapter "Structure" in
Rheinwerk Computing C. Every Chapter from now on will be executed
in an other Function. The User can call every function by typing
the right letter in the Structure, like t 0 (Example function
Hello World).