Function to read out which category of file it is
PROBLEMS:
in the tutorial we are able to enter as much files as desired.
Now we face again that argc and **argv are not possible to input
them as easy as shown. So we should have to implement annother logic
to read multible files. For that one should implement a seperation
rule read out the input string until seperator and then take that
part to analysis. Repead that for any numbers of seperator.